feat: format price/discount/cash_received with thousands separator via input-rupiah
The finance edit money inputs used type=number, which shows no thousands
separator while typing and displayed raw decimal-cast strings like
25000.000 for new rows. Replaced with the x-input-rupiah component
(decimals=3, same as cashier): display uses id-ID format (25.000 /
1.234,567), raw value is kept for the hidden submission fields, per-row
price syncs via update-rupiah-value on mount and when row.price changes.
Product pick normalizes price to a number, and edit() now casts product
prices to float so @json emits clean numbers.

+++ app/Http/Controllers/FinanceReportController.php
@@ -400,7 +400,7 @@ class FinanceReportController extends Controller
             return [
                 'id' => $p->id,
                 'name' => $p->name,
-                'price' => $firstStock?->price_consument ?? 0,
+                'price' => (float) ($firstStock?->price_consument ?? 0),
                 'stock' => $stockSum,
             ];
         });
